Add jinja2 tags
The tags are also included in the basic comment flavour package.
It literally just imports the stock jinja2tags, but allows a developer
to import everything from the basic app. Other comment flavours in the
future will likely need custom tags for custom functionality, so this is
for consistency with future flavours.